cqlrt and cqlrt_cf had much lower warning settings (#163)
* signed changes

* removing assorted workings with stricter compilation

* correct for cqlrt_cf as well

* Update for cqlrt_cf

* Update sources/cqlrt_common.h

* Update sources/cqlrt.h

* Update sources/cqlrt_common.c

+++ sources/cqlrt_cf/cqlrt_cf.c
@@ -40,7 +40,7 @@ char *_Nullable cql_copy_string_to_stack_or_heap(
           cstr = stack_string;
       }
       else {
-        heap_string = malloc(size);
+        heap_string = malloc((size_t)size);
         if (CFStringGetCString(cf_string_ref, heap_string, size, kCFStringEncodingUTF8)) {
           cstr = heap_string;
         }
